+++ docs/Changelog.md
@@ -12,6 +12,7 @@ sending a mail to <afl-users+subscribe@googlegroups.com>.
     - new llvm pass: dict2file via AFL_LLVM_DICT2FILE, create afl-fuzz -x
       dictionary of string comparisons found during compilation
     - not overriding -Ox or -fno-unroll-loops anymore
+    - allow up to 4 -x command line options
 
 
 ### Version ++2.67d (develop)